🔨🤖 keep email schemas out of site bundle
Infer request contracts from shared schemas exposed through a server-only entrypoint. Trust D1 subscriber rows and update the JavaScript and CSS budgets for the new newsletter UI.
