+++ .bundlemonrc.json
@@ -4,12 +4,12 @@
     "files": [
         {
             "path": "owid.mjs",
-            "maxSize": "800kB",
+            "maxSize": "805kB",
             "maxPercentIncrease": 2
         },
         {
             "path": "owid.css",
-            "maxSize": "62kB"
+            "maxSize": "64kB"
         }
     ],
     "groups": [
