+++ packages/@ourworldindata/types/package.json
@@ -5,6 +5,10 @@
     "type": "module",
     "main": "src/index.ts",
     "types": "src/index.ts",
+    "exports": {
+        ".": "./src/index.ts",
+        "./email-notifications-schemas": "./src/EmailNotificationsSchemas.ts"
+    },
     "directories": {
         "lib": "src"
     },
