+++ packages/@ourworldindata/types/src/EmailNotificationsTypes.ts
@@ -1,5 +1,12 @@
-import * as z from "zod/mini"
-import { LatestType } from "./domainTypes/Latest.js"
+import type { LatestType } from "./domainTypes/Latest.js"
+import type { EmailNotificationsPreferences } from "./EmailNotificationsSchemas.js"
+
+export type {
+    EmailNotificationsPreferences,
+    EmailNotificationsRequestLinkRequest,
+    EmailNotificationsSubscribeRequest,
+    EmailNotificationsUpdatePreferencesRequest,
+} from "./EmailNotificationsSchemas.js"
 
 export const EMAIL_NOTIFICATIONS_FREQUENCIES = ["daily", "weekly"] as const
 
@@ -75,34 +82,6 @@ export const EMAIL_NOTIFICATIONS_CONTENT_TYPE_LABELS: Record<
     announcement: "Announcements",
 }
 
-// Mirrors the validation in the email-notifications subscribe Cloudflare
-// Function (see functions/api/email-notifications/subscribe.ts), which is the
-// authoritative place where requests are validated.
-export const EmailNotificationsPreferencesTypeObject = z.object({
-    // Topic tag names from the topic tag graph. Articles, data insights
-    // and data updates are only sent when they match one of these topics;
-    // the "announcement" content type is topic-independent. An empty array
-    // means "all topics" — stored that way so topic areas added later are
-    // automatically included. (The subscribe form still makes users pick at
-    // least one pill and translates a full selection to []; that rule is
-    // client-side only, since the server can't tell "nothing selected" from
-    // "all topics".)
-    topicTags: z
-        .array(z.string().check(z.minLength(1), z.maxLength(100)))
-        .check(z.maxLength(64)),
-    contentTypes: z
-        .array(z.enum(EMAIL_NOTIFICATIONS_CONTENT_TYPES))
-        .check(
-            z.minLength(1),
-            z.maxLength(EMAIL_NOTIFICATIONS_CONTENT_TYPES.length)
-        ),
-    frequency: z.enum(EMAIL_NOTIFICATIONS_FREQUENCIES),
-})
-
-export type EmailNotificationsPreferences = z.infer<
-    typeof EmailNotificationsPreferencesTypeObject
->
-
 /**
  * Union of two sets of preferences, used when the public subscribe form is
  * submitted for an address that already exists: the form is tokenless, so it
@@ -130,82 +109,10 @@ export function mergeEmailNotificationsPreferences(
     }
 }
 
-export const EmailNotificationsSubscribeRequestTypeObject = z
-    .object({
-        email: z.email().check(z.maxLength(254)),
-        // Preferences for the new notifications system, stored in our own
-        // database. Omitted if the user only wants the OWID Brief.
-        notifications: z.optional(EmailNotificationsPreferencesTypeObject),
-        // The OWID Brief newsletter stays in Mailchimp.
-        subscribeToOwidBrief: z.boolean(),
-    })
-    .check(
-        z.refine(
-            (request) =>
-                request.notifications !== undefined ||
-                request.subscribeToOwidBrief,
-            "Select email notifications or the OWID Brief newsletter"
-        )
-    )
-
-export type EmailNotificationsSubscribeRequest = z.infer<
-    typeof EmailNotificationsSubscribeRequestTypeObject
->
-
 export interface EmailNotificationsSubscribeResponse {
     ok?: boolean
     error?: string
 }
-
-// Request a magic link for updating preferences. Either an email address
-// (from the enter-email UI; unknown addresses get the identical response and
-// no email — see the request-link function) or a token: the permanent
-// per-user token from an email footer link, or an expired magic-link token
-// (its resend button).
-export const EmailNotificationsRequestLinkRequestTypeObject = z
-    .object({
-        email: z.optional(z.email().check(z.maxLength(254))),
-        token: z.optional(z.string().check(z.minLength(1), z.maxLength(100))),
-    })
-    .check(
-        z.refine(
-            (request) => Boolean(request.email) !== Boolean(request.token),
-            "Provide either an email or a token"
-        )
-    )
-
-export type EmailNotificationsRequestLinkRequest = z.infer<
-    typeof EmailNotificationsRequestLinkRequestTypeObject
->
-
-// Save from the magic-link preferences page. The magic link itself was the
-// proof of inbox control, so changes apply immediately (no second
-// confirmation email). Notifications based on these preferences and the
-// Mailchimp-owned OWID Brief are independent subscriptions.
-// `subscribeToOwidBrief` is omitted when Mailchimp's current state could not
-// be loaded and its toggle is unavailable.
-const EmailNotificationsUpdatePreferencesCommonShape = {
-    token: z.string().check(z.minLength(1), z.maxLength(100)),
-    subscribeToOwidBrief: z.optional(z.boolean()),
-}
-
-export const EmailNotificationsUpdatePreferencesRequestTypeObject =
-    z.discriminatedUnion("subscribeToTopicNotifications", [
-        z.object({
-            ...EmailNotificationsUpdatePreferencesCommonShape,
-            subscribeToTopicNotifications: z.literal(true),
-            preferences: EmailNotificationsPreferencesTypeObject,
-        }),
-        z.object({
-            ...EmailNotificationsUpdatePreferencesCommonShape,
-            subscribeToTopicNotifications: z.literal(false),
-            preferences: z.optional(z.never()),
-        }),
-    ])
-
-export type EmailNotificationsUpdatePreferencesRequest = z.infer<
-    typeof EmailNotificationsUpdatePreferencesRequestTypeObject
->
 
 export interface EmailNotificationsPreferencesResponse {
     email?: string
